app/pyproject.toml
[tool.poetry]
name = "online-events-backend"
version = "0.1.0"
description = ""
authors = ["Khaled Farhat <khaled.farhat.cpp@gmail.com>"]
readme = "README.md"
packages = [{include = "core"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-environ = "^0.10.0"
pillow = "^9.4.0"
django-rest-knox = "^4.2.0"
drf-spectacular = "^0.26.1"
django-cors-headers = "^4.0.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
channels = {extras = ["daphne"], version = "^4.0.0"}
daphne = "^4.0.0"
channels-redis = "^4.1.0"
celery = "^5.3.1"
redis = "^5.0.0"


[tool.poetry.group.dev.dependencies]
django-silk = "^5.0.3"
pytest-django = "^4.5.2"
factory-boy = "^3.3.0"
black = "^23.7.0"
flake8 = "^6.1.0"
pytest-mock = "^3.11.1"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
extend-exclude = '''
/(
  | migrations
)/
'''
